// src/generated/theporndb/client/types.gen.ts
export type RequestResult<
  TData = unknown,
  TError = unknown,
  ThrowOnError extends boolean = boolean
> = ThrowOnError extends true
  ? Promise<{
      data: TData extends Record<string, unknown> ? TData[keyof TData] : TData
      response: Response
    }>
  : Promise<
      (
        | {
            data: TData extends Record<string, unknown> ? TData[keyof TData] : TData
            error: undefined
          }
        | {
            data: undefined
            error: TError extends Record<string, unknown> ? TError[keyof TError] : TError
          }
      ) & {
        response: Response
      }
    >