// src/generated/theporndb/types.gen.ts
export type DirectorResource = {
  id?: number
  uuid?: string
  name?: string
  slug?: string
}

/**
 * A resource that contains information about an image
 */
export type MediaResource = {
  id?: number
  url?: string
  size?: number
  order?: number
}

/**
 * Contains information about a performer
 */
export type PerformerResource = {
  id?: string
  _id?: number
  slug?: string
  name?: string
  disambiguation?: string
  bio?: string | null
  rating?: number
  is_parent?: boolean
  extras?: {
    gender?: string | null
    birthday?: string | null
    birthday_timestamp?: number | null
    deathday?: string | null
    deathday_timestamp?: number | null
    birthplace?: string | null
    birthplace_code?: string | null
    astrology?: string | null
    ethnicity?: string | null
    nationality?: string | null
    hair_colour?: string | null
    eye_colour?: string | null
    weight?: string | null
    height?: string | null
    measurements?: string | null
    cupsize?: string | null
    tattoos?: string | null
    piercings?: string | null
    waist?: string | null
    hips?: string | null
    career_start_year?: number | null
    career_end_year?: number | null
    fake_boobs?: boolean
    same_sex_only?: boolean
    links?: {
      [key: string]: string | null
    } | null
  }
  aliases?: Array<string> | null
  image?: string
  thumbnail?: string
  face?: string
  posters?: Array<MediaResource> | null
  site_performers?: Array<PerformerSiteResource> | null
}
// ...
export type MarkerResource = {
  id?: number
  title?: string
  start_time?: number
  end_time?: number | null
  created_at?: string
}

export type SceneResource = {
  id?: string
  _id?: number
  title?: string
  type?: string
  slug?: string
  external_id?: string
  description?: string
  rating?: number
  site_id?: number
  date?: string
  url?: string
  image?: string
  back_image?: string
  poster_image?: string
  poster?: string
  trailer?: string
  duration?: number
  format?: string
  sku?: string
  background?: {
    full?: string
    large?: string
    medium?: string
    small?: string
  }
  media?: MediaResource
  created?: string
  last_updated?: string
  performers?: Array<PerformerResource>
  site?: SiteResource
  tags?: Array<TagResource>
  hashes?: Array<SceneHashBasicResponse>
  markers?: Array<MarkerResource>
  directors?: Array<DirectorResource>
  /**
   * If this entity is a movie, contains scenes attached
   */
  scenes?: Array<SceneResource>
  /**
   * If this entity is a scene, contains movies this scene appears in
   */
  movies?: Array<SceneResource>
  links?: {
    [key: string]: string | null
  } | null
}

export type SiteResource = {
  uuid?: string
  id?: number
  parent_id?: number
  network_id?: number
  name?: string
  short_name?: string
  url?: string
  description?: string
  rating?: number
  logo?: string
  favicon?: string
  poster?: string
  network?: SiteResource
  parent?: SiteResource
}

export type TagResource = {
  id?: number
  uuid?: string
  name?: string
  parents?: Array<TagResource>
}

/**
 * A resource that contains information about a user
 */
export type UserResource = {
  id?: number
  name?: string
}

export type SceneHashBasicResponse = {
  can_delete?: boolean | null
  created_at?: string
  duration?: number
  hash?: string
  id?: number
  scene_id?: number
  submissions?: number
  type?: SceneHashEnum
  updated_at?: string
}

export type SceneCardResponse = {
  background?: unknown
  date?: string | null
  default_background?: string | null
  duration?: number | null
  id?: number | null
  is_collected?: boolean | null
  is_hidden?: boolean | null
  link?: string | null
  performers?: unknown
  site?: unknown
  slug?: string | null
  title?: string | null
  type?: SceneTypeEnum
}

/**
 * Pagination meta data
 */
export type Meta = {
  total?: number
  per_page?: number
  current_page?: number
  last?: number
}
// ...
export type GetPerformerScenesResponses = {
  /**
   * List Performer Scenes
   */
  200: {
    /**
     * An array of SceneResources
     */
    data?: Array<SceneResource>
    meta?: Meta
  }
}